done fixing design issues
